+++ src/tile/TileManager.java
@@ -97,8 +97,8 @@ public class TileManager {
 
             int worldX = worldCol * gp.tileSize;
             int worldY = worldRow * gp.tileSize;
-            int screenX = worldX - gp.player.worldX + gp.player.screenX;
-            int screenY = worldY - gp.player.worldY + gp.player.screenY;
+            int screenX = worldX - (int)gp.player.worldX + gp.player.screenX;
+            int screenY = worldY - (int)gp.player.worldY + gp.player.screenY;
 
             if(worldX > gp.player.worldX - gp.player.screenX - gp.tileSize &&
                worldX < gp.player.worldX + gp.screenWidth - gp.player.screenX &&
